+++ NEWS.md
@@ -1,4 +1,4 @@
-# tabularaster dev
+# tabularaster 0.3.0
 
 * bufext made defunct